requirements.txt
absl-py==2.2.2
accelerate==0.19.0
addict==2.4.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
basicsr==1.4.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
datasets==3.0.0
DateTime==5.5
diffusers==0.12.1
dill==0.3.8
docker-pycreds==0.4.0
facexlib==0.3.0
filelock==3.16.1
filterpy==1.4.5
fonttools==4.56.0
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
gfpgan==1.3.8
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.71.0
huggingface-hub==0.25.1
idna==3.10
imageio==2.37.0
importlib_metadata==8.5.0
Jinja2==3.1.4
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.44.0
lmdb==1.6.2
Markdown==3.8
MarkupSafe==2.1.5
matplotlib==3.10.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.3
numba==0.61.2
numpy==2.1.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==10.4.0
piq==0.8.0
platformdirs==4.3.6
protobuf==5.28.2
psutil==6.0.0
pyarrow==17.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
realesrgan==0.3.0
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-image==0.25.2
scipy==1.15.2
sentry-sdk==2.17.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.13.3
tb-nightly==2.20.0a20250503
tensorboard-data-server==0.7.2
tifffile==2025.3.30
tokenizers==0.13.3
tomli==2.2.1
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.28.1
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wandb==0.18.3
Werkzeug==3.1.3
xxhash==3.5.0
yapf==0.43.0
yarl==1.12.1
zipp==3.20.2
zope.interface==7.1.0
